requirements.txt
python-dotenv==1.0.0
pyPDF2==3.0.1
faiss-cpu==1.7.4
gradio==4.12.0
requests==2.31.0
python-docx==0.8.11